Clean map-api backend routes
This commit moves the endpoints to the routes folder. This makes the code cleaner.
This also removes the html pages for adding data from the backend.
Adding beacons, booths, events can still be done via curl.
Listed backend APIs in README.
