+++ containers/map-api/package.json
@@ -15,12 +15,13 @@
     "url": "https://github.com/IBM/secret-map-dashboard"
   },
   "dependencies": {
+    "assert": "^1.4.x",
     "body-parser": "^1.17.x",
-    "mongoose": "^4.13.x",
+    "cors": "^2.8.x",
     "express": "^4.15.x",
     "fs": "~0.0.1",
-    "assert": "^1.4.x",
-    "cors": "^2.8.x"
+    "mongoose": "^4.13.x",
+    "router": "^1.3.2"
   },
   "author": "Anthony Amanse",
   "license": "Apache-2.0"
