--- containers/map-api/public/addBeacon.html
@@ -1,19 +1,0 @@
-<html>
-  <head>
-    <title>Add Beacon</title>
-    <link rel="stylesheet" href="css/style.css">
-  </head>
-  <body>
-    <a href="/">Home</a> / <a href="/booth">Add Booth</a> / <a href="/beacon">Add Beacon</a> / <a href="/event">Add Event</a>
-    / <a href="/booth/list">Booth List</a> / <a href="/beacon/list">Beacon List</a> / <a href="/event/list">Event List</a>
-    <p>Hello, Beacon!</p>
-    <input id="beaconId" name="beaconId" placeholder="Beacon ID" >
-    <input id="x" type="number" name="x" placeholder="X">
-    <input id="y" type="number" name="y" placeholder="Y">
-    <input id="minCount" type="number" name="minCount" placeholder="Minimum Count">
-    <input id="maxCount" type="number" name="maxCount" placeholder="Maximum Count">
-    <button id="addBeacon">Add Beacon</button>
-
-    <script src="js/submit.js"></script>
-  </body>
-</html>
