--- containers/map-api/public/addBooth.html
@@ -1,23 +1,0 @@
-<html>
-  <head>
-    <title>Add Booth</title>
-    <link rel="stylesheet" href="css/style.css">
-  </head>
-  <body>
-    <a href="/">Home</a> / <a href="/booth">Add Booth</a> / <a href="/beacon">Add Beacon</a> / <a href="/event">Add Event</a>
-    / <a href="/booth/list">Booth List</a> / <a href="/beacon/list">Beacon List</a> / <a href="/event/list">Event List</a>
-    <p>Hello, Booth!</p>
-    <input id="boothId" name="boothId" placeholder="Booth ID" >
-    <input id="unit" name="unit" placeholder="Unit Name">
-    <input id="description" name="description" placeholder="Description">
-    <input id="measurementUnit" name="measurementUnit" placeholder="Measurement Unit">
-    <input id="width" type="number" name="width" placeholder="Width">
-    <input id="height" type="number" name="height" placeholder="Height">
-    <input id="x" type="number" name="x" placeholder="X Location">
-    <input id="y" type="number" name="y" placeholder="Y Location">
-    <input id="contact" name="contact" placeholder="Contact">
-    <button id="addBooth">Add Booth</button>
-
-    <script src="js/submit.js"></script>
-  </body>
-</html>
