--- containers/map-api/public/addEvent.html
@@ -1,21 +1,0 @@
-<html>
-  <head>
-    <title>Add Event</title>
-    <link rel="stylesheet" href="css/style.css">
-  </head>
-  <body>
-    <a href="/">Home</a> / <a href="/booth">Add Booth</a> / <a href="/beacon">Add Beacon</a> / <a href="/event">Add Event</a>
-    / <a href="/booth/list">Booth List</a> / <a href="/beacon/list">Beacon List</a> / <a href="/event/list">Event List</a>
-    <p>Hello, Event!</p>
-    <input id="eventId" name="eventId" placeholder="Event ID" >
-    <input id="eventName" name="eventName" placeholder="Event Name">
-    <input id="location" name="location" placeholder="Location">
-    <input id="startDate" type="date" name="startDate" placeholder="Start Date">
-    <input id="endDate" type="date" name="endDate" placeholder="End Date">
-    <input id="map" name="map" placeholder="Map">
-    <input id="beacons" name="beacons" placeholder="Beacons">
-    <button id="addEvent">Add Event</button>
-
-    <script src="js/submit.js"></script>
-  </body>
-</html>
